Re-Factoring:
1 - Changed mapping methods in CustomerDetailsController to return 404 status when id's do not exist.

2 - Added test class CustomerDetailsControllerTest to test endpoints.

3 - Re-factored modal handler methods in admin.js

+++ src/main/java/com/example/solobackend/controller/CustomerDetailsController.java
@@ -2,7 +2,9 @@ package com.example.solobackend.controller;
 
 import com.example.solobackend.model.CustomerDetails;
 import com.example.solobackend.service.CustomerDetailsService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,6 +14,7 @@ import java.util.Optional;
 public class CustomerDetailsController {
 
     private final CustomerDetailsService service;
+    public static final String ID_NOT_FOUND_ERROR_MSG = "Driver Not Found for ID: ";
 
     public CustomerDetailsController(CustomerDetailsService service) {
         this.service = service;
@@ -24,7 +27,13 @@ public class CustomerDetailsController {
 
     @GetMapping("/customerDetails")
     Optional<CustomerDetails> getCustomerById(@RequestParam("id") Long id) {
-        return service.getCustomer(id);
+        Optional<CustomerDetails> customerToGet = service.getCustomer(id);
+        if(customerToGet.isPresent()){
+            return customerToGet;
+        }else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, ID_NOT_FOUND_ERROR_MSG + id);
+        }
     }
 
     @PostMapping("/customerDetails")
@@ -34,11 +43,25 @@ public class CustomerDetailsController {
 
     @DeleteMapping("customerDetails")
     void delete(@RequestParam("id") Long id) {
-        service.deleteCustomer(id);
+        Optional<CustomerDetails> customerToDelete = service.getCustomer(id);
+        if(customerToDelete.isPresent()){
+            service.deleteCustomer(id);
+            return;
+        }else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, ID_NOT_FOUND_ERROR_MSG + id);
+        }
     }
 
     @PutMapping("customerDetails")
     void update(@RequestParam("id") Long id, @RequestParam("newTelephoneNumber") String newTelephoneNumber) {
-        service.updateCustomer(id, newTelephoneNumber);
+        Optional<CustomerDetails> customerToUpdate = service.getCustomer(id);
+        if(customerToUpdate.isPresent()){
+            service.updateCustomer(id, newTelephoneNumber);
+            return;
+        }else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, ID_NOT_FOUND_ERROR_MSG + id);
+        }
     }
 }
